requirements.txt
playwright>=1.39.0
schedule>=1.2.1
requests>=2.31.0
pyyaml>=6.0
openai>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.5.0
